Added SoundManager.js
The SoundManager Class is honestly a big mess. I'm trying to wait until code works and looks a lot better before I upload it but here it is anyway! The purpose of this Class is to Manager all the Sound Instances created by CreateJS. When you play a "Talkie" sound it will be added to a Talkie Layer, playing a Song will be added to the Music Layer and so forth.