Added modified signals.js
ScummJS uses JS Signals <http://millermedeiros.github.com/js-signals/> with a few modifications. In all honesty this may be removed in favor of CreateJS's EventDispatcher Class. Using Signals was a left over idea from the ActionScript 3 version of this project. Right now Signals are used for custom events and 90% of the time only one listener will be attached to an event...Signals seem to be meant for events with multiple listeners.