Added PuttCostume_migrate_to_children.js
This branch will feature the first Costume Class file. The reason the name features "Putt" is because a lot of things were hardcoded to directly enable the Putt-Putt Costume. This file will directly show the move away from that so there will only need to be one file for all. As well, the file will gain a new name "Actor.js" as this will be created by the ActorController class. 

The biggest reason it's in the _layersRemoval branch is due to the decision of wanting to remove the "_layers" variable form the top level of the class. Where "actors[1]._layers._layer1" will now be more Createjs friendly doing "actors[1].COSTUME.getChildAt(0)"