Optomized Actor.js
Some of the major changes to the Actor class was how the layers are handled. I am trying to completely remove the need of the _layers variable. Having the VO is actually helpful in many cases and to protect legacy demos the _layer variable can still be set but it has been deprecated for CreateJS's getChildAt Functions