Costume Layer JSON
There are two ways of giving a Costume layers, one is by having the game load in a json file. In this json shows the structure of a Costume with layers. The layers it self can have the complete Sprite Sheet in it, as show by this example's "body" layer. A "Small Object" where certain layer definers and a Sprite Sheet json source can live, demonstrated by the "head" layer. Or simply just a String pointer to a Sprite Sheet json, as seen in the "eyes-normal" and "eyes-player" layer. The complete Sprite Sheet or "Small Object" can have definers like "isHidden" or "canColorize", respectfully one will make the layer hidden on adding to the Stage/Canvas and the other will allow the Sprite to have multi-color variants of it's images.